requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
jinja2
python-multipart
asyncpg
psycopg2-binary
alembic
python-dotenv
greenlet
cloudinary